pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "ultra_ray"
version = "1.0.0"
description = "Full-Path Ray Tracing in Physics-Based Ultrasound Simulation"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Felix Duelmer", email = "felix.duelmer@tum.de"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "scipy>=1.6.0",
    "matplotlib",
    "ultraspy",
    "configargparse",
    "trimesh",
    "pyglet<2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]
cuda = [
    "cupy-cuda12x",  # for CUDA 12.x systems
]
cuda11 = [
    "cupy-cuda11x",  # for CUDA 11.x systems
]

[project.urls]
Homepage = "https://github.com/Felixduelmer/ultra_ray"
Repository = "https://github.com/Felixduelmer/ultra_ray"
Documentation = "https://github.com/Felixduelmer/ultra_ray#readme"
"Bug Reports" = "https://github.com/Felixduelmer/ultra_ray/issues"